Add sparse reward model

+++ src/rl_final_project/q_functions.py
@@ -19,7 +19,7 @@ class QTabular:
     """Discretized tabular Q state-action-value function."""
 
     def __init__(
-        self, n_actions: int, n_feat: int, discrete_scale: int = 40
+        self, n_actions: int, n_feat: int, discrete_scale: float = 40
     ) -> None:
         """Define a tabular Q function.
